+++ docs/CHANGELOG.md
@@ -40,6 +40,7 @@ VoxEdit:
    - Fixed palette panel issue if you changed the order of colors
    - Added more templates to the new-scene-dialog
    - Made the new-scene-dialog scrollable
+   - Allow to change the transform of a node with updating the children
 
 Thumbnailer: